requirements.txt
pymavlink
pika
pyyaml